SPA rework
