+++ package.json
@@ -27,6 +27,7 @@
   },
   "dependencies": {
     "@blockly/zoom-to-fit": "^5.0.12",
+    "bootstrap": "^5.3.3",
     "dotenv": "^16.4.1",
     "ejs": "^3.1.9"
   }
