+++ public/index.html
@@ -140,9 +140,7 @@
     
     <script src="./js/bootstrap.min.js"></script>
     <script type="module" src="./client.js"></script>
-    <script type="module" src="./js/utils.js"></script>
     <script type="module" src="./js/login.js"></script>
-    <script type="module" src="./js/home.js"></script>
 </body>
 
 </html>
