+++ rollup.config.dist.mjs
@@ -19,7 +19,7 @@ export default [
         output: {
             file: './public/client.js',
             name: 'client',
-            format: 'iife',
+            format: 'es',
             sourcemap: true
         },
