+++ src/Game/Phaser/LevelPlayer/Classes/ArticodingSprite.ts
