+++ src/Game/Phaser/LevelPlayer/Classes/TrapObject.ts
